requirements.txt
azure-cognitiveservices-vision-customvision==3.1.0
azure-common==1.1.28
azure-core==1.26.4
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.6.2
urllib3==2.0.2
